Move wtgs_tools class to util; remove obsolete methods
