+++ src/lib/djerba/plugins/wgts/snv_indel/tools.py
@@ -14,11 +14,10 @@ import seaborn as sns
 import numpy as np
 import logging
 import djerba.core.constants as core_constants
-import djerba.plugins.wgts.common.cnv.constants as cnv_constants
+import djerba.plugins.wgts.cnv_purple.legacy_constants as cnv_constants
 import djerba.plugins.wgts.snv_indel.constants as sic
 from djerba.mergers.gene_information_merger.factory import factory as gim_factory
 from djerba.mergers.treatment_options_merger.factory import factory as tom_factory
-from djerba.plugins.wgts.common.tools import wgts_tools
 from djerba.util.environment import directory_finder
 from djerba.util.html import html_builder
 from djerba.util.image_to_base64 import converter
@@ -27,6 +26,7 @@ from djerba.util.oncokb.annotator import annotator_factory
 from djerba.util.oncokb.tools import levels as oncokb_levels
 from djerba.util.oncokb.tools import gene_summary_reader
 from djerba.util.subprocess_runner import subprocess_runner
+from djerba.util.wgts.tools import wgts_tools
 
 class whizbam:
